refactor: 서비스 로직 클래스 분리
